refactor: convert unban.js to ts
